Move search box to hero bar
