+++ dashboard/origin-mlx/src/pages/MetaFeaturedPage/MetaFeaturedPage.tsx
@@ -13,19 +13,50 @@
 *  See the License for the specific language governing permissions and 
 *  limitations under the License. 
 */ 
-import React, { useContext, ComponentProps, useEffect } from 'react'
+import React, { useContext, useState, ComponentProps, useEffect } from 'react'
 import StoreContext from '../../lib/stores/context'
 import { fetchArtifact } from '../../lib/api/artifacts';
 import { Artifact, FETCH_ARTIFACT_ASSETS } from '../../lib/stores/artifacts'
 import { SET_ACTIVE_PAGE } from '../../lib/stores/pages';
 import { getUserInfo, hasRole, canShow } from '../../lib/util'
+import fuzzysort from 'fuzzysort'
 
 import Button from '../../components/Button'
 import Hero from '../../components/Hero';
 import Icon from '@material-ui/core/Icon'
 import Link from '../../components/Link'
 import MetaFeatured from './MetaFeatured';
+
+import IconButton from "@material-ui/core/IconButton";
+import InputAdornment from "@material-ui/core/InputAdornment";
+import SearchIcon from "@material-ui/icons/Search";
+import FormControl from '@material-ui/core/FormControl';
+import Input from '@material-ui/core/Input';
+
 import '../../styles/Models.css';
+import { withStyles } from '@material-ui/core/styles';
+
+const CssInput = withStyles({
+  root: {
+    '& label.Mui-focused': {
+      color: 'green',
+    },
+    '& .MuiInput-underline:after': {
+      borderBottomColor: 'green',
+    },
+    '& .MuiOutlinedInput-root': {
+      '& fieldset': {
+        borderColor: 'red',
+      },
+      '&:hover fieldset': {
+        borderColor: 'yellow',
+      },
+      '&.Mui-focused fieldset': {
+        borderColor: 'green',
+      },
+    },
+  },
+})(Input);
 
 
 export interface MetaFeaturedPageProps extends ComponentProps<any> {
@@ -56,6 +87,19 @@ function MetaFeaturedPage(props: MetaFeaturedPageProps) {
   const { store, dispatch } = useContext(StoreContext)
   const { artifacts, settings } = store
   const assets: Artifact[] = artifacts[assetType]
+  const [search, setSearch] = useState('')
+
+  let filteredAssets = assets
+  const assetNames = assets.filter(asset => asset.featured && asset.publish_approved).map((asset: any) => asset.name)
+  if (search !== '') {
+    const searchResults = fuzzysort.go(search, assetNames)
+    filteredAssets = searchResults.reduce(function(result, searchResult) {
+      const index = assets.findIndex((asset: any) => asset.name === searchResult.target)
+      if (index !== -1)
+        result.push(assets[index])
+      return result;
+    }, []);
+  }
 
   const API = settings.endpoints.api.value || settings.endpoints.api.default
   const upload = settings.capabilities.upload
@@ -107,8 +151,8 @@ function MetaFeaturedPage(props: MetaFeaturedPageProps) {
         }
         <Link to="https://github.com/machine-learning-exchange/mlx">
           <Button
-            className="hero-buttons-outline"
-            variant="outlined"
+            className="hero-buttons"
+            variant="contained"
             color="primary"
           >
             Github
@@ -126,13 +170,37 @@ function MetaFeaturedPage(props: MetaFeaturedPageProps) {
             </Button>
           </Link>
         }
+        <FormControl className="search-box">
+          <CssInput
+            id="standard-adornment-weight"
+            value={search}
+            onChange={(event: React.ChangeEvent<HTMLInputElement>)=>setSearch(event.target.value)}
+            placeholder="Search"
+            endAdornment={
+              <InputAdornment position="end">
+                <IconButton>
+                  <SearchIcon style={{fill: "#fff"}} />
+                </IconButton>
+              </InputAdornment>
+            }
+            inputProps={{
+              endAdornment: (
+                <InputAdornment position="end">
+                  <IconButton>
+                    <SearchIcon />
+                  </IconButton>
+                </InputAdornment>
+              )
+            }}
+          />
+        </FormControl>
       </Hero>
         {children}
         {assets && 
           <MetaFeatured 
             numFeatured={numFeatured}
             assetType={assetType}
-            assets={assets.filter(asset => asset.featured && asset.publish_approved)} 
+            assets={filteredAssets} 
           /> 
         }
     </div>
